+++ packages/l10n_currencies/lib/src/data/mr_currencies_l10n.data.dart
@@ -159,7 +159,7 @@ class MrCurrenciesL10N extends IsoLocaleMapper<String> {
       "VUV": "वानुआतु वाटु",
       "WST": "सामोअन टाला",
       "XAF": "मध्य आफ्रिकन [CFA] फ्रँक",
-      "XCD": "पूर्व कॅरीबियन डॉलर",
+      "XCD": "पूर्व कॅरिबियन डॉलर",
       "XCG": "कॅरिबियन गिल्डर",
       "XOF": "पश्चिम आफ्रिकन [CFA] फ्रँक",
       "XPF": "[CFP] फ्रँक",